+++ src/main/java/roomescape/domain/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
